requirements.txt
motor
pydantic
python-dotenv
pymongo[srv]